moves the game engine logic into the right place
