+++ src/server/server/net_entities/game.js
@@ -1,5 +1,6 @@
 const EventEmitter = require('events');
 const Wall = require('../2d/wall.js');
+const Pong = require('../engines/pong.js');
 
 class Game extends EventEmitter{
   constructor(con) {
@@ -7,7 +8,6 @@ class Game extends EventEmitter{
 
     con.on('close', this.shutdown.bind(this))
 
-    this.winning_score = 5;
     this.active_tokens = ['foo'];
     this.con = con;
     this.game_running = false;
@@ -19,69 +19,35 @@ class Game extends EventEmitter{
     this.intervals = [];
     this.game_running = false;
 
-    //this.walls = [];
-    //this.walls << new Wall({x: 0, y: 0}, {x: 1080, y: 0});
-    //this.walls << new Wall({x: 0, y: 1080}, {x: 1080, y: 1080});
-    //this.walls << new Wall({x: 0, y: 0}, {x: 0, y: 1080});
-    //this.walls << new Wall({x: 1080, y: 0}, {x: 1080, y: 1080});
-
     this.generate_token();
     this.intervals.push(setInterval(this.generate_token.bind(this), (1000 * 60)));
 
-    this.gamefield = {
-      width: 1080,
-      height: 1920,
-      paddle_size: 230,
-      paddle_offset: 50
-    };
-
     this.players = [
       {
-        pos: 400,
-        name: 'Bot',
-        score: 0,
         controller: null
       },
       {
-        pos: 400,
-        name: 'Bot',
-        score: 0,
         controller: null
       }
     ];
-
-    this.ball = {
-      x: this.gamefield.width / 2,
-      y: this.gamefield.height / 2,
-      radius: 35,
-      vx: 10,
-      vy: 8
-    };
-
-    this.ball_paused = false;
   }
 
   reset() {
+    this.pong.cleanup();
+    this.pong = null;
     this._clear_intervals();
     this.construct_field();
   }
 
   start() {
-    this.game_running = true;
+    this.pong = new Pong(this.players);
 
-    this.intervals.push(setInterval(this.send_game_state.bind(this), (1000 / 34)));
-    this.intervals.push(setInterval(this.game_tick.bind(this), (1000 / 60)));
+    this.pong.on('game_state', this.send_game_state.bind(this));
+    this.pong.on('game_over', this.game_over.bind(this));
 
     this.send_to(0, 'game_start');
     this.send_to(1, 'game_start');
     this.con.send('game_start');
-
-    console.log('new game');
-
-    this.ball_paused = true;
-    setTimeout(() => {
-      this.ball_paused = false;
-    }, 3000);
   }
 
   generate_token() {
@@ -95,99 +61,12 @@ class Game extends EventEmitter{
     this.send_token();
   }
 
-  game_tick() {
-    let bfx = this.ball.x + this.ball.vx;
-    let bfy = this.ball.y + this.ball.vy;
-    let ball_radius = this.ball.radius;
-    let invert_x = false;
-    let invert_y = false;
-
-    // Process player movement
-    let player_speed = 22;
-
-    this.players.forEach((player) => {
-      let controller = player.controller;
-
-      if(controller) {
-        if(controller.get_key_state('left')) {
-          if(player.pos - player_speed > 0) player.pos -= player_speed;
-          else player.pos = 0;
-        }
-
-        if(controller.get_key_state('right')) {
-          if(player.pos + player_speed + this.gamefield.paddle_size < this.gamefield.width) player.pos += player_speed;
-          else player.pos = this.gamefield.width - this.gamefield.paddle_size;
-        }
-      }
-    });
-
-    // Paddle hit detection
-    let in_p1, in_p2;
-
-    let player = this.players[0];
-    let distance_to_paddle = (bfy - ball_radius) - this.gamefield.paddle_offset;
-
-    in_p1 = bfy < (this.gamefield.paddle_offset + ball_radius);
-    in_p1 = in_p1 && distance_to_paddle > -Math.abs(this.ball.vx);
-    in_p1 = in_p1 && (bfx + ball_radius) > player.pos;
-    in_p1 = in_p1 && bfx < (player.pos + this.gamefield.paddle_size);
-
-    player = this.players[1];
-    distance_to_paddle = -((bfy + ball_radius) - (this.gamefield.height - this.gamefield.paddle_offset));
-
-    in_p2 = bfy > (this.gamefield.height - this.gamefield.paddle_offset - ball_radius);
-    in_p2 = in_p2 && distance_to_paddle > -Math.abs(this.ball.vx);
-    in_p2 = in_p2 && (bfx + ball_radius) > player.pos;
-    in_p2 = in_p2 && bfx < (player.pos + this.gamefield.paddle_size);
-
-    if(in_p1 || in_p2) {
-      invert_y = true;
-    }
-
-    //this.walls.forEach((wall) => {
-    //  let collision = wall.check_collision(this.ball, this.ball.radius);
-    //});
-
-    // Wall Detection
-    let top_wall = -(ball_radius * 2);
-    let bottom_wall = this.gamefield.height + (ball_radius * 2);
-
-    if(bfx + ball_radius > this.gamefield.width || (bfx - ball_radius) < 0) invert_x = true;
-    if(bfy + ball_radius > bottom_wall || bfy < top_wall) {
-      if(bfy + ball_radius > this.gamefield.height) this.player_scored(0);
-      if(bfy < top_wall) this.player_scored(1);
-
-      invert_y = true;
-    }
-
-    if(invert_x) this.ball.vx *= -1;
-    if(invert_y) this.ball.vy *= -1;
-
-    // Process ball movement
-    if(!this.ball_paused) {
-      this.ball.x += this.ball.vx;
-      this.ball.y += this.ball.vy;
-    }
-  }
-
-  player_scored(id) {
-    this.ball.x = this.gamefield.width / 2;
-    this.ball.y = this.gamefield.height / 2;
-
-    this.players[id].score++;
-
-    if(this.players[id].score == this.winning_score) {
-      console.log(`GAME OVER player ${id+1} won!!!`);
-      this.send_to(0, 'game_over', {id: id});
-      this.send_to(1, 'game_over', {id: id});
-      this.con.send('game_over', {id: id});
-      this.reset();
-    } else {
-      this.ball_paused = true;
-      setTimeout(() => {
-        this.ball_paused = false;
-      }, 2000);
-    }
+  game_over(id) {
+    console.log(`GAME OVER player ${id+1} won!!!`);
+    this.send_to(0, 'game_over', {id: id});
+    this.send_to(1, 'game_over', {id: id});
+    this.con.send('game_over', {id: id});
+    this.reset();
   }
 
   send_token() {
@@ -198,28 +77,7 @@ class Game extends EventEmitter{
     this.con.send('token', data);
   }
 
-  send_game_state() {
-    let p1 = this.players[0];
-    let p2 = this.players[1];
-    let current_token = this.active_tokens[0];
-
-    let data = {
-      current_join_token: current_token,
-      p1: {
-        pos: p1.pos,
-        score: p1.score,
-        name: p1.name
-      },
-      p2: {
-        pos: p2.pos,
-        score: p2.score,
-        name: p2.name
-      }
-    };
-
-    data.ball = this.ball;
-    data.gamefield = this.gamefield;
-
+  send_game_state(data) {
     this.con.send('game_state', data);
   }
